+++ gzap/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- New:
+  - Compiled logic and tuning data is now cached in your AP directory. This uses
+    a few tens of MB but also makes apworld loading about 4x faster after the
+    first time.
 - Fix:
   - The item icon mapper now properly detects FreeDoom 1/2 as having Doom-
     compatible spritesheets.